training/requirements.txt
numpy<2
pandas
scikit-learn
pyyaml
tqdm
matplotlib
pytest
torch
onnx
onnxruntime
